Renamed config file
