+++ src/report/js/ReportConfig.js
